requirements.txt
python-dotenv
rich
discord.py
pynacl
